Distinction between format and human readable
