+++ src/yabc/formats/adhoc.py
@@ -49,7 +49,8 @@ class AdhocParser(Format):
     This class translates CSV rows into transaction objects.
     """
 
-    EXCHANGE_NAME = "adhoc"
+    FORMAT_NAME = "adhoc CSV"
+    EXCHANGE_HUMAN_READABLE_NAME = "adhoc"
 
     def validate_headers(self, curr):
         for header_name in _FIELD_NAMES:
@@ -153,7 +154,7 @@ def _handle_spending(date, curr):
         symbol_received="USD",
         fees=fees,
         fee_symbol=fee_coin,
-        source=AdhocParser.EXCHANGE_NAME,
+        source=AdhocParser.exchange_name(),
     )
 
 
@@ -173,7 +174,7 @@ def _handle_buy(date, curr):
         quantity_received=curr[_RECEIVED_AMOUNT],
         fees=fees,
         fee_symbol=feecoin,
-        source=AdhocParser.EXCHANGE_NAME,
+        source=AdhocParser.exchange_name(),
     )
 
 
@@ -192,7 +193,7 @@ def _handle_gift_received(date, curr):
         # What was actually received
         quantity_received=curr[_RECEIVED_AMOUNT],
         symbol_received=curr[_RECEIVED_CURRENCY],
-        source=AdhocParser.EXCHANGE_NAME,
+        source=AdhocParser.exchange_name(),
     )
 
 
@@ -204,7 +205,7 @@ def _handle_gift_sent(date, curr):
         quantity_traded=curr[_TRADED_AMOUNT],
         symbol_received="USD",
         quantity_received=0,
-        source=AdhocParser.EXCHANGE_NAME,
+        source=AdhocParser.exchange_name(),
     )
 
 
@@ -224,7 +225,7 @@ def _handle_mining(date, curr):
         symbol_traded="USD",
         quantity_traded=mined_value,
         date=date,
-        source=AdhocParser.EXCHANGE_NAME,
+        source=AdhocParser.exchange_name(),
     )
     return trans
 
@@ -253,7 +254,7 @@ def _make_adhoc_sell(date, curr):
             date=date,
             fees=fee,
             fee_symbol=fee_coin,
-            source=AdhocParser.EXCHANGE_NAME,
+            source=AdhocParser.exchange_name(),
         )
