+++ src/yabc/formats/coinbase.py
@@ -55,7 +55,8 @@ def txs_from_coinbase(f):
 
 
 class CoinbaseParser(Format):
-    EXCHANGE_NAME = "Coinbase (Trades Format)"
+    FORMAT_NAME = "Coinbase (Trades Format)"
+    EXCHANGE_HUMAN_READABLE_NAME = "Coinbase"
 
     def __init__(self, file_or_fname):
         self._file = file_or_fname
